Created checks to ensure long enough keys
Added code that checks if the entered plain text and the key are the same length. If they are not the same length, a key value is taken from the original and appended to the end. This ensured that keys are *always* long enough to complete vigenere encryption.

Modified:     src/lightman210567_encipher/encipher.py